Moving shared certification reducers to new file
